create shared component for title
