+++ components/d2l-activity-editor/d2l-activity-content-editor/module/d2l-activity-content-module-detail.js
@@ -1,12 +1,10 @@
-import 'd2l-inputs/d2l-input-text.js';
-import 'd2l-tooltip/d2l-tooltip';
+import '../shared-components/d2l-activity-content-editor-title.js';
 import '../../d2l-activity-html-editor';
-import '@brightspace-ui/core/components/button/button-subtle.js';
 import { AsyncContainerMixin, asyncStates } from '@brightspace-ui/core/mixins/async-container/async-container-mixin.js';
 import { css, html } from 'lit-element/lit-element.js';
-import { DEBOUNCE_TIMEOUT, TITLE_MAX_LENGTH } from '../d2l-activity-content-editor-detail.js';
 import { ActivityEditorMixin } from '../../mixins/d2l-activity-editor-mixin.js';
 import { ContentModuleEntity } from 'siren-sdk/src/activities/content/ContentModuleEntity.js';
+import { DEBOUNCE_TIMEOUT } from '../d2l-activity-content-editor-detail.js';
 import { Debouncer } from '@polymer/polymer/lib/utils/debounce.js';
 import { EntityMixinLit } from 'siren-sdk/src/mixin/entity-mixin-lit.js';
 import { ErrorHandlingMixin } from '../../error-handling-mixin.js';
@@ -19,12 +17,6 @@ import { SkeletonMixin } from '@brightspace-ui/core/components/skeleton/skeleton
 import { timeOut } from '@polymer/polymer/lib/utils/async.js';
 
 class ContentModuleDetail extends AsyncContainerMixin(SkeletonMixin(ErrorHandlingMixin(LocalizeActivityEditorMixin(EntityMixinLit(RtlMixin(ActivityEditorMixin(MobxLitElement))))))) {
-
-	static get properties() {
-		return {
-			_titleError: { type: String }
-		};
-	}
 
 	static get styles() {
 		return  [
@@ -55,35 +47,26 @@ class ContentModuleDetail extends AsyncContainerMixin(SkeletonMixin(ErrorHandlin
 		this.saveOrder = 2000;
 	}
 
+	connectedCallback() {
+		super.connectedCallback();
+		this.saveTitle = this.saveTitle.bind(this);
+	}
+
 	render() {
 		const moduleEntity = moduleStore.getContentModuleActivity(this.href);
-		let title = '';
 		let descriptionRichText = undefined;
 		this.skeleton = true;
 		if (moduleEntity) {
 			// Show loading skeleton until we have the module entity loaded into state
 			this.skeleton = false;
-			title = moduleEntity.title;
 			descriptionRichText = moduleEntity.descriptionRichText;
 		}
 
 		return html`
-			<div id="content-title-container">
-				<d2l-input-text
-					id="content-title"
-					maxlength="${TITLE_MAX_LENGTH}"
-					value="${title}"
-					@change="${this._saveOnChange('title')}"
-					@input="${this._saveTitleOnInput}"
-					label="${this.localize('content.name')} *"
-					aria-invalid="${this._titleError ? 'true' : ''}"
-					prevent-submit
-					novalidate
-					?skeleton="${this.skeleton}"
-				>
-				</d2l-input-text>
-				${this._renderTitleTooltip()}
-			</div>
+			<d2l-activity-content-editor-title
+				.entity=${moduleEntity}
+				.saveFn=${this.saveTitle}
+			></d2l-activity-content-editor-title>
 			<slot name="dueDate"></slot>
 			<div id="content-description-container">
 				<div class="d2l-activity-label-container d2l-label-text d2l-skeletize">
@@ -135,6 +118,14 @@ class ContentModuleDetail extends AsyncContainerMixin(SkeletonMixin(ErrorHandlin
 		await moduleEntity.save();
 	}
 
+	saveTitle(value) {
+		const moduleEntity = moduleStore.getContentModuleActivity(this.href);
+		if (!moduleEntity) {
+			return;
+		}
+		moduleEntity.setTitle(value);
+	}
+
 	_onRichtextChange(e) {
 		const descriptionRichText = e.detail.content;
 		this._debounceJobs.description = Debouncer.debounce(
@@ -166,34 +157,6 @@ class ContentModuleDetail extends AsyncContainerMixin(SkeletonMixin(ErrorHandlin
 			return;
 		}
 		moduleEntity.setDescription(richText);
-	}
-
-	_saveOnChange(jobName) {
-		this._debounceJobs[jobName] && this._debounceJobs[jobName].flush();
-	}
-
-	_saveTitle(value) {
-		const moduleEntity = moduleStore.getContentModuleActivity(this.href);
-		if (!moduleEntity) {
-			return;
-		}
-		moduleEntity.setTitle(value);
-	}
-
-	_saveTitleOnInput(e) {
-		const title = e.target.value;
-		const isTitleEmpty = (title || '').trim().length === 0;
-
-		if (isTitleEmpty) {
-			this.setError('_titleError', 'content.emptyNameField', 'title-tooltip');
-		} else {
-			this.clearError('_titleError');
-			this._debounceJobs.title = Debouncer.debounce(
-				this._debounceJobs.title,
-				timeOut.after(DEBOUNCE_TIMEOUT),
-				() => this._saveTitle(title)
-			);
-		}
 	}
 }
 customElements.define('d2l-activity-content-module-detail', ContentModuleDetail);
